Update: cloudinary added

+++ backend/controllers/productController.js
@@ -16,10 +16,16 @@ const addProduct = async (req, res) => {
     const images = [image1, image2, image3, image4].filter((item) => item !== undefined)
 
     // images upload cloudinary
-    let imageUrl = await Promise.all(
+    let imageData = await Promise.all(
       images.map(async (item) => {
-        let result = await cloudinary.uploader.upload(item.path, { resource_type: 'image' });
-        return result.secure_url
+        let result = await cloudinary.uploader.upload(item.path, {
+          resource_type: 'image',
+          folder: 'ecommerce_app'
+        });
+        return {
+          url: result.secure_url,
+          public_id: result.public_id
+        }
       })
     )
 
@@ -32,7 +38,7 @@ const addProduct = async (req, res) => {
       price: Number(price),
       bestseller: bestseller === "true" ? true : false,
       sizes: JSON.parse(sizes),
-      image: imageUrl,
+      image: imageData,
       date: Date.now()
     };
 
@@ -58,6 +64,17 @@ const listProducts = async (req, res) => {
 // fuction for removing product
 const removeProduct = async (req, res) => {
   try {
+
+    const product = await productModel.findById(req.body.id);
+
+    if (!product) {
+      return res.json({ success: false, message: "Product not found" });
+    }
+
+    for (let img of product.image) {
+      await cloudinary.uploader.destroy(img.public_id);
+    }
+
     await productModel.findByIdAndDelete(req.body.id)
     res.json({ success: true, message: "Product removed" })
   } catch (error) {
